Moved dependencies to deps folder
