+++ src/tile/TileDatabase.h
@@ -2,7 +2,7 @@
 #define NONAME_TILEDATABASE_H
 
 #include <SFML/Graphics/Texture.hpp>
-#include "../utils/json.hpp"
+#include "../../deps/json.h"
 
 #include "Tile.h"
